switch to the higher level safe API
