+++ Cargo.toml
@@ -7,9 +7,6 @@ version = "0.1.0"
 particle-hal = { git = "https://github.com/japaric/particle-hal" }
 photon = { git = "https://github.com/japaric/photon" }
 
-[dev-dependencies]
-numtoa = "0.0.7"
-
 [profile.dev]
 debug = false
 debug-assertions = false
