+++ examples/variable.rs
@@ -1,4 +1,4 @@
-//! Example for the Cloud variable API
+//! Example for the Particle.variable API
 
 #![no_std]
 
@@ -6,45 +6,34 @@ extern crate particle_hal as hal;
 #[macro_use]
 extern crate photon;
 
-use core::ptr;
-
-use hal::{PinMode, Spark_Data_TypeDef, pin_t};
-
-const D7: pin_t = 7;
+use hal::{LED, PinMode, cloud};
 
 app! {
     setup: setup,
     loop: loop_,
 }
 
+const PERIOD: u32 = 1_000;
 static mut COUNTER: i32 = 0;
 
 fn setup() {
-    unsafe {
-        static NAME: &[u8] = b"counter\0";
+    LED.pin_mode(PinMode::Output);
 
-        hal::HAL_Pin_Mode(D7, PinMode::OUTPUT);
-
-        if hal::spark_variable(
-            NAME.as_ptr() as *const i8,
-            &COUNTER as *const _ as *const _,
-            Spark_Data_TypeDef::CLOUD_VAR_INT,
-            ptr::null_mut(),
-        )
-        {
-            hal::HAL_GPIO_Write(D7, 1);
-        } else {
-            hal::HAL_GPIO_Write(D7, 0);
-        }
+    if cloud::variable("counter", unsafe { &COUNTER }).is_ok() {
+        LED.high()
+    } else {
+        LED.low()
     }
 }
 
 fn loop_() {
+    LED.high();
+    hal::sleep_ms(PERIOD / 2);
+
+    LED.low();
+    hal::sleep_ms(PERIOD / 2);
+
     unsafe {
-        hal::HAL_GPIO_Write(D7, 0);
-        hal::HAL_Delay_Milliseconds(500);
-        hal::HAL_GPIO_Write(D7, 1);
-        hal::HAL_Delay_Milliseconds(500);
         COUNTER += 1;
     }
 }
